v1.0.1
Se cambió el comando generate para poder enviar un id de canal y se genere el embed con un JSON que se pase.
Se agregó el sistema de verificación.

+++ src/commands/Generate.ts
@@ -1,5 +1,6 @@
-import { Message, MessageEmbed } from "discord.js";
+import { GuildEmoji, Message, MessageEmbed, MessageEmbedOptions, TextChannel } from "discord.js";
 import Command from "./Command";
+const emojiRegExp = /<a?:[^:]+:(\d+)>/gm;
 
 export class Generate extends Command {
     readonly name = "generate";
@@ -7,19 +8,37 @@ export class Generate extends Command {
     readonly args = true;
     readonly permission = "ADMINISTRATOR";
 
-    executed(message: Message, emojiStr: string | undefined, ...args: string[]): void {
-        const emojiRegExp = /<a?:[^:]+:(\d+)>/gm;
-        const emoji = emojiStr ? message.guild?.emojis.cache.get(emojiRegExp.exec(emojiStr)![1]) : undefined;
-        const title = args.length > 0 ? args.join(" ") : "Creación de pedido";
-        message.delete();
+    async executed(message: Message, emojiStr: string, canalStr: string, ...args: string[]): Promise<any> {
+        const emojiID = emojiRegExp.exec(emojiStr);
+        let emoji: GuildEmoji | string | undefined = emojiStr ? message.guild?.emojis.cache.get(emojiID && emojiID[1] ? emojiID[1] : "") : undefined;
+        if (message.deletable) message.delete();
 
-        message.channel
-            .send(
-                new MessageEmbed()
-                    .setTitle(title)
-                    .setDescription(`Para crear una conversación sobre tu pedido reacciona aquí. ${emoji || "📩"}`)
-                    .setColor(0x6666ff)
-            )
-            .then((msg) => msg.react(emoji || "📩"));
+        let canal: TextChannel | undefined | null = message.channel as TextChannel;
+        if (!emoji) emoji = "📩";
+        if (!isNaN(Number(canalStr))) canal = message.client.channels.cache.get(canalStr) as TextChannel | undefined;
+        else args = [canalStr, ...args];
+
+        if (!canal && !isNaN(Number(canalStr)))
+            try {
+                canal = (await message.client.channels.fetch(canalStr)) as TextChannel;
+            } catch (_) {
+                null;
+            }
+        if (!canal) return message.channel.send("`❌|` Canal no encontrado.");
+
+        let mensajeEmbed: MessageEmbed | MessageEmbedOptions | undefined;
+        try {
+            mensajeEmbed = JSON.parse(args.join(" "));
+        } catch (error) {
+            return message.channel.send("`❌|` Esctructura no válida.");
+        }
+
+        return canal
+            .send(new MessageEmbed(mensajeEmbed))
+            .then((msg) => {
+                msg.react(emoji!);
+                if (message.channel.id != canal!.id) message.channel.send("`✅|` Mensaje enviado al canal.");
+            })
+            .catch((error: Error) => message.channel.send(`\`❌|\` Error al mandar el mensaje: \`\`\`${error.stack || error.message}\`\`\``));
     }
 }
